Join the two different line charts into one file to maintain a standar.
