+++ medTracker/View/Analysis/Chart/ChartView.swift
@@ -37,14 +37,10 @@ struct ChartView: View {
             
             // Switch between line and bar graph. Pass the currentTab selected to switch time-zones.
             if (isLineGraph) {
-                if (symptom.isQuantitative) {
-                    LineChartView_Cuant(symptom: symptom, registers: registers, currentTab: $currentTab)
-                } else {
-                    LineChartView_Cual(symptom: symptom, registers: registers, currentTab: $currentTab)
-                }
+                LineChartView(symptom: symptom, registers: registers, currentTab: $currentTab)
             } else {
                 if (symptom.isQuantitative) {
-                    BarChart2(symptom: symptom, registers: registers, currentTab: $currentTab)
+                    BarChartView_Cuant(symptom: symptom, registers: registers, currentTab: $currentTab)
                 } else {
                     BarChartView_Cual(symptom: symptom, registers: registers, currentTab: $currentTab)
                 }
