+++ medTracker/View/Analysis/Chart/LineChartView.swift
@@ -1,16 +1,16 @@
 //
-//  LineChartView_Cual.swift
+//  LineChartView.swift
 //  medTracker
 //
-//  Created by Joel Alejandro Milla Lopez on 23/03/24.
+//  Created by Joel Alejandro Milla Lopez on 22/03/24.
 //
 
 import SwiftUI
 import Charts
 
-struct LineChartView_Cual: View {
+struct LineChartView: View {
     // Variables to show information
-    @State var symptom: Symptom
+    let symptom: Symptom
     @ObservedObject var registers: RegisterList
     // MARK: View Properties
     @Binding var currentTab: String
@@ -23,13 +23,13 @@ struct LineChartView_Cual: View {
         // MARK: Chart that changes when the currentTab (time zone selected changes)
         AnimatedCharts()
             .onChange(of: currentTab) { newValue in
-                filteredRegisters = registers.registers[symptom.id.uuidString]?.filterBy(currentTab) ?? []
+                filterRegisters()
                 // Re-Animating View
                 animateGraph(fromChange: true)
             }
         // This onChange will trigger the graph to update when a new register is created
             .onChange(of: registers.registers) { _ in
-                filteredRegisters = registers.registers[symptom.id.uuidString]?.filterBy(currentTab) ?? []
+                filterRegisters()
                 // Re-Animating View
                 animateGraph()
             }
@@ -37,9 +37,13 @@ struct LineChartView_Cual: View {
     
     @ViewBuilder
     func AnimatedCharts() -> some View {
+        // Max value to extend the y-scale of the data
+        let max = filteredRegisters.max { item1, item2 in
+            return item2.amount > item1.amount
+        }?.amount ?? 0
         // Values to block the x scale from moving
-        let minDate = filteredRegisters.map { $0.date }.min() ?? Date()
-        let maxDate = filteredRegisters.map { $0.date }.max() ?? Date()
+        let minDate = filteredRegisters.first?.date ?? Date.now
+        let maxDate = filteredRegisters.last?.date ?? Date.now
         
         Chart {
             ForEach(filteredRegisters) { register in
@@ -75,69 +79,52 @@ struct LineChartView_Cual: View {
                 if let currentActiveItem, currentActiveItem.id.uuidString == register.id.uuidString {
                     // Add a rule on the x value on the graph
                     RuleMark(
-                        x: .value("Fecha", currentActiveItem.date)
+                        x: .value("Fecha", register.date)
+                        //                        yStart: .value("Min", 0),
+                        //                        yEnd: .value("Cantidad", currentActiveItem.cantidad)
                     )
                     // Add an annotation on top of the vertical line to show the value of the nearest item
                     .annotation(position: .top) {
-                        VStack(alignment: .leading, spacing: 6) {
-                            // Show the date of the current value and the value
-                            Text(currentActiveItem.date.dateToStringMDH())
-                                .font(.caption)
-                                .foregroundStyle(.gray)
-                            // Obtain the image of the current value selected and show it
-                            let imageName = HelperFunctions.getImage(of: currentActiveItem.amount)
-                            HStack {
-                                Spacer()
-                                Image(imageName)
-                                    .resizable()
-                                    .scaledToFit()
-                                    .frame(width: 30)
-                                Spacer()
-                            }
-                        }
-                        .padding(.horizontal, 10)
-                        .padding(.vertical, 4)
-                        .background {
-                            RoundedRectangle(cornerRadius: 6, style: .continuous)
-                                .fill(.white.shadow(.drop(radius: 2)))
-                        }
-                        // Move the annotation when it is on the corners so the annotation shows clearly and not on borders
-                        .offset(x: currentActiveItem.date.dateToStringMDH() == minDate.dateToStringMDH() ? 35 : currentActiveItem.date.dateToStringMDH() == maxDate.dateToStringMDH() ? -20 : 0)
+                        Annotation(symptom: symptom, register: currentActiveItem, minDate: minDate, maxDate: maxDate)
                     }
                 }
             }
         }
         // MARK: Customizing x and y axis length
         .chartXScale(domain: minDate...maxDate)
-        .chartYScale(domain: [0, 150]) // bigger number, smaller the bar charts
-        // MARK: Customizing the x labels and y labels
+        .chartYScale(domain: 0...(1.5 * max)) // bigger number, smaller the bar charts
+        // Add the y axis labels when the symptom is Quantitaitve
         .chartYAxis {
-            AxisMarks(preset: .aligned, position: .trailing, values: .stride(by: 50)) { value in
-                if let yValue = value.as(Double.self) {
-                    AxisGridLine() // Show the grid line on the y axis
-                    AxisValueLabel {
-                        // On values 0-50-100 show images instead of numbers
-                        switch yValue {
-                        case 0:
-                            Image("sadder_face")
-                                .resizable()
-                                .scaledToFit()
-                                .frame(width: 30)
-                        case 50:
-                            Image("sad_face")
-                                .resizable()
-                                .scaledToFit()
-                                .frame(width: 30)
-                        case 100:
-                            Image("happier_face")
-                                .resizable()
-                                .scaledToFit()
-                                .frame(width: 30)
-                        default:
-                            Text("")
+            if (!symptom.isQuantitative) {
+                AxisMarks(preset: .aligned, position: .trailing, values: .stride(by: 50)) { value in
+                    if let yValue = value.as(Double.self) {
+                        AxisGridLine() // Show the grid line on the y axis
+                        AxisValueLabel {
+                            // On values 0-50-100 show images instead of numbers
+                            switch yValue {
+                            case 0:
+                                Image("sadder_face")
+                                    .resizable()
+                                    .scaledToFit()
+                                    .frame(width: 30)
+                            case 50:
+                                Image("sad_face")
+                                    .resizable()
+                                    .scaledToFit()
+                                    .frame(width: 30)
+                            case 100:
+                                Image("happier_face")
+                                    .resizable()
+                                    .scaledToFit()
+                                    .frame(width: 30)
+                            default:
+                                Text("")
+                            }
                         }
                     }
                 }
+            } else {
+                AxisMarks() // Default behavior for quantitative data
             }
         }
         // MARK: Gesture to highlight current bar.
@@ -170,10 +157,14 @@ struct LineChartView_Cual: View {
         })
         .frame(height: 250)
         .onAppear {
-            // Filter the current regiser based on the current time zone.
-            filteredRegisters = registers.registers[symptom.id.uuidString]?.filterBy(currentTab) ?? []
+            filterRegisters()
             animateGraph()
         }
+    }
+    
+    // Function to update the filteredRegisters
+    func filterRegisters() {
+        filteredRegisters = registers.registers[symptom.id.uuidString]?.filterBy(currentTab) ?? []
     }
     
     // MARK: Function to animate the graph
@@ -187,17 +178,61 @@ struct LineChartView_Cual: View {
     }
 }
 
+//MARK: View to show an annotation depending if the symptom is quant or qual
+struct Annotation: View {
+    let symptom: Symptom
+    let register: Register
+    
+    let minDate: Date
+    let maxDate: Date
+    
+    var body: some View {
+        VStack(alignment: .leading, spacing: 6) {
+            // Show the date of the current value and the value
+            Group {
+                if (symptom.isQuantitative) {
+                    Text(register.date.dateToStringMDH())
+                        .font(.caption)
+                        .foregroundStyle(.gray)
+                    Text(register.amount.asString())
+                        .font(.title3.bold())
+                } else {
+                    Text(register.date.dateToStringMDH())
+                        .font(.caption)
+                        .foregroundStyle(.gray)
+                    // Obtain the image of the current value selected and show it
+                    let imageName = HelperFunctions.getImage(of: register.amount)
+                    HStack {
+                        Spacer()
+                        Image(imageName)
+                            .resizable()
+                            .scaledToFit()
+                            .frame(width: 30)
+                        Spacer()
+                    }
+                }
+            }
+        }
+        .padding(.horizontal, 10)
+        .padding(.vertical, 4)
+        .background {
+            RoundedRectangle(cornerRadius: 6, style: .continuous)
+                .fill(.white.shadow(.drop(radius: 2)))
+        }
+        // Move the annotation when it is on the corners so the annotation shows clearly and not on borders
+        .offset(x: register.date.dateToStringMDH() == minDate.dateToStringMDH() ? 35 : register.date.dateToStringMDH() == maxDate.dateToStringMDH() ? -20 : 0)
+    }
+}
+
 #Preview {
     NavigationStack {
         @State var repository = Repository(user: User(id: "3zPDb70ofQQHximl1NXwPMgIhMR2", rol: "Paciente", email: "joel@mail.com", phone: "", name: "Joel", clinicalHistory: "", sex: "", birthdate: Date.now, height: "", doctors: ["doc@mail.com"]))
         @State var registers: RegisterList = RegisterList(repository: repository)
         
-        @State var symptom = Symptom(name: "", icon: "heart", description: "", isQuantitative: true, units: "kg", isActive: true, color: "#000000", notification: "")
+        let symptom = Symptom(name: "", icon: "heart", description: "", isQuantitative: true, units: "kg", isActive: true, color: "#000000", notification: "")
         
         @State var currentTab: String = "Semana"
         
-        LineChartView_Cual(symptom: symptom, registers: registers, currentTab: $currentTab)
+        LineChartView(symptom: symptom, registers: registers, currentTab: $currentTab)
     }
 }
-
-
